requirements.txt
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==1.7.3
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.6
Django==4.1.2
django-browser-reload==1.6.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-haystack==3.2.1
django-import-export==3.0.2
django-redis==5.2.0
django-redis-admin==0.2.0
django-tailwind==3.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
filelock==3.8.0
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.1.0
platformdirs==2.5.4
pluggy==0.3.1
poyo==0.5.0
psycopg2==2.9.4
py==1.11.0
pycountry==22.3.5
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.6
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
text-unidecode==1.3
tox==2.1.1
tzdata==2022.5
urllib3==1.26.12
virtualenv==20.16.7
whitenoise==6.2.0
Whoosh==2.7.4
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
